Added after_login function to app/views.py. 1. resp arg passed to after_login contains info returned by OpenID provider. 2. After email is validated, the database is searched for the email provided. If not found, we add a new user to the DB. We also handle the case of a missing nickname here, as some OpenID providers don't give this info. 3. We then load remember_me from the Flask session object, if it is available. (It was stored in login view function). 4. login_user is called to register a valid login. 5. Lastly, we redirect to the 'next' page or the 'index' page if 'next' was not provided in the request.
Flask-Login protects views against non-logged inusers by adding the login_required decorator. The protected URL cannot be accessed by a non-logged in user, and they are redirected to the login page.
Flask-Login stores the original URL as the 'next' page. We return the user to this page once the login process is over.